sandbox/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
docker>=7.0.0
python-dateutil>=2.9.0
pydantic>=2.0.0
